requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-filter==25.1
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.7
tzdata==2025.1
